# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GamePanel API - Game Server & VPS Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /nodes:
    get:
      operationId: listNodes
      tags: [nodes]
      summary: List all nodes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
    post:
      operationId: createNode
      tags: [nodes]
      summary: Create a node
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeBody"
      responses:
        "201":
          description: Created node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
  /nodes/{id}:
    get:
      operationId: getNode
      tags: [nodes]
      summary: Get a node
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
    put:
      operationId: updateNode
      tags: [nodes]
      summary: Update a node
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeBody"
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
    delete:
      operationId: deleteNode
      tags: [nodes]
      summary: Delete a node
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /nodes/{id}/stats:
    get:
      operationId: getNodeStats
      tags: [nodes]
      summary: Get node resource stats
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Node stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeStats"
  /nodes/{id}/regenerate-token:
    post:
      operationId: regenerateNodeToken
      tags: [nodes]
      summary: Regenerate node API token
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeTokenResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required:
        - email
        - username
        - password
        - firstName
        - lastName
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        serverLimit:
          type: integer
        vpsLimit:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - firstName
        - lastName
        - role
        - isActive
        - serverLimit
        - vpsLimit
        - createdAt
...
    Node:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        fqdn:
          type: string
        ip:
          type: string
        port:
          type: integer
        isOnline:
          type: boolean
        cpuCores:
          type: integer
        memoryMb:
          type: integer
        diskGb:
          type: integer
        location:
          type: string
        description:
          type: string
        apiToken:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - fqdn
        - ip
        - port
        - isOnline
        - cpuCores
        - memoryMb
        - diskGb
        - createdAt
    CreateNodeBody:
      type: object
      properties:
        name:
          type: string
        fqdn:
          type: string
        ip:
          type: string
        port:
          type: integer
        cpuCores:
          type: integer
        memoryMb:
          type: integer
        diskGb:
          type: integer
        location:
          type: string
        description:
          type: string
      required:
        - name
        - fqdn
        - ip
        - port
        - cpuCores
        - memoryMb
        - diskGb
    UpdateNodeBody:
      type: object
      properties:
        name:
          type: string
        fqdn:
          type: string
        ip:
          type: string
        port:
          type: integer
        cpuCores:
          type: integer
        memoryMb:
          type: integer
        diskGb:
          type: integer
        location:
          type: string
        description:
          type: string
    NodeStats:
      type: object
      properties:
        cpuUsage:
          type: number
        memoryUsed:
          type: integer
        memoryTotal:
          type: integer
        diskUsed:
          type: integer
        diskTotal:
          type: integer
        networkIn:
          type: number
        networkOut:
          type: number
        serverCount:
          type: integer
        vpsCount:
          type: integer
      required:
        - cpuUsage
        - memoryUsed
        - memoryTotal
        - diskUsed
        - diskTotal
        - networkIn
        - networkOut
        - serverCount
        - vpsCount
    NodeTokenResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
...